Chore: #18 공통으로 사용되는 chakra 관련 스타일 파일 생성 및 정리
