+++ FE/issue-tracker/src/components/common/chakraStyle.ts
@@ -1,0 +1,21 @@
+const menuBtnStyle = {
+  marginBottom: '16px',
+  display: 'flex',
+  width: '100%',
+  height: '100%',
+  variant: 'unstyled',
+  justifyContent: 'space-between',
+  borderRadius: '16px',
+};
+
+const menuItemStyle = {
+  display: 'flex',
+  justifyContent: 'space-between',
+};
+
+const checkBoxStyle = {
+  colorScheme: 'cyan',
+  borderColor: 'gr_label',
+};
+
+export { checkBoxStyle, menuItemStyle, menuBtnStyle };
