--- FE/issue-tracker/src/components/issues/table/style.ts
@@ -1,8 +1,0 @@
-export const menuBtnStyle = {
-  background: 'transparent',
-  padding: '5px',
-  width: 'auto',
-  height: '30px',
-  color: 'gr_label',
-  marginLeft: '12px',
-};
