--- FE/issue-tracker/src/components/newIssue/chakraStyle.ts
@@ -1,46 +1,0 @@
-const titleInput = {
-  marginBottom: '16px',
-  padding: '16px',
-  height: '56px',
-  variant: 'filled',
-  background: 'gr_inputBackground',
-  placeholder: '제목',
-  borderRadius: '14px',
-};
-
-const contentsInput = {
-  size: 'md',
-  height: '291px',
-  padding: '16px 24px',
-  background: 'gr_inputBackground',
-  placeholder: '코멘트를 입력하세요',
-  variant: 'filled',
-};
-
-const menuBtnStyle = {
-  padding: '34px 32px',
-  display: 'flex',
-  width: '100%',
-  height: '96px',
-  background: 'gr_offWhite',
-  variant: 'unstyled',
-  justifyContent: 'space-between',
-};
-
-const menuItemStyle = {
-  display: 'flex',
-  justifyContent: 'space-between',
-};
-
-const checkBoxStyle = {
-  colorScheme: 'cyan',
-  borderColor: 'gr_label',
-};
-
-export {
-  titleInput,
-  contentsInput,
-  checkBoxStyle,
-  menuItemStyle,
-  menuBtnStyle,
-};
