+++ FE/issue-tracker/src/components/newIssue/style.ts
@@ -1,0 +1,27 @@
+const titleInput = {
+  marginBottom: '16px',
+  padding: '16px',
+  height: '56px',
+  variant: 'filled',
+  background: 'gr_inputBackground',
+  placeholder: '제목',
+  borderRadius: '14px',
+};
+
+const contentsInput = {
+  size: 'md',
+  height: '291px',
+  padding: '16px 24px',
+  background: 'gr_inputBackground',
+  placeholder: '코멘트를 입력하세요',
+  variant: 'filled',
+  borderRadius: '16px',
+};
+
+const progressBar = {
+  hasStripe: true,
+  colorScheme: 'green',
+  borderRadius: '16px',
+};
+
+export { titleInput, contentsInput, progressBar };
